+++ lib/clevertap_plugin.dart
@@ -4,6 +4,8 @@ import 'package:flutter/services.dart';
 
 typedef void CleverTapInAppNotificationDismissedHandler(
     Map<String, dynamic> mapList);
+typedef void CleverTapInAppNotificationShowHandler(
+    Map<String, dynamic> map);
 typedef void CleverTapInAppNotificationButtonClickedHandler(
     Map<String, dynamic>? mapList);
 typedef void CleverTapProfileDidInitializeHandler();
@@ -27,6 +29,8 @@ typedef void CleverTapPushPermissionResponseReceivedHandler(bool accepted);
 class CleverTapPlugin {
   late CleverTapInAppNotificationDismissedHandler
       cleverTapInAppNotificationDismissedHandler;
+  late CleverTapInAppNotificationShowHandler
+      cleverTapInAppNotificationShowHandler;
   late CleverTapInAppNotificationButtonClickedHandler
       cleverTapInAppNotificationButtonClickedHandler;
   late CleverTapProfileDidInitializeHandler
@@ -70,6 +74,11 @@ class CleverTapPlugin {
       case "inAppNotificationDismissed":
         Map<dynamic, dynamic> args = call.arguments;
         cleverTapInAppNotificationDismissedHandler(
+            args.cast<String, dynamic>());
+        break;
+      case "inAppNotificationShow":
+        Map<dynamic, dynamic> args = call.arguments;
+        cleverTapInAppNotificationShowHandler(
             args.cast<String, dynamic>());
         break;
       case "onInAppButtonClick":
@@ -136,6 +145,11 @@ class CleverTapPlugin {
   void setCleverTapInAppNotificationDismissedHandler(
           CleverTapInAppNotificationDismissedHandler handler) =>
       cleverTapInAppNotificationDismissedHandler = handler;
+
+  /// Only for Android - Define a method to handle inApp notification shown
+  void setCleverTapInAppNotificationShowHandler(
+          CleverTapInAppNotificationShowHandler handler) =>
+      cleverTapInAppNotificationShowHandler = handler;
 
   /// Define a method to handle inApp notification button clicked
   void setCleverTapInAppNotificationButtonClickedHandler(
@@ -828,5 +842,10 @@ class CleverTapPlugin {
   ///Returns true if push permission is enabled.
   static Future<bool?> getPushNotificationPermissionStatus() async {
     return await _channel.invokeMethod('getPushNotificationPermissionStatus', {});
+  }
+
+  ///Only for Android - Unregisters PushPermissionNotificationResponseListener
+  static Future<void> unregisterPushPermissionNotificationResponseListener() async {
+    return await _channel.invokeMethod('unregisterPushPermissionNotificationResponseListener', {});
   }
 }